Add files via upload
The optimization model developed using Python, maximizing the population (used as a proxy for fire incidences) and ensuring 100% coverage for the city of Magdeburg by Professional Fire Services in Scenario 2a. This file also contains the sensitivity analysis where we are varying the budget constraint in terms of maximum number of fire stations to be opened and analyzing the effect of this variation on population covered by the fire services.