Population coverage w.r.t budget constraint
The optimization model developed using Python, maximizing population (used as a proxy for fire incidences) coverage for the city of Magdeburg by Professional Fire Services. Budget constraints in terms of maximum number of stations to be opened has been introduced as an additional constraint in the model. Also sensitivity analysis has been performed by varying the number of fire stations to be opened and analyzing the effect on population coverage.